dom/workers/test/serviceworkers/test_installation_simple.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<head>
  <title>Bug 930348 - test stub Navigator ServiceWorker utilities.</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test"></pre>
<script class="testbody" type="text/javascript">

  function simpleRegister() {
    var p = navigator.serviceWorker.register("worker.js", { scope: "simpleregister/*" });
    ok(p instanceof Promise, "register() should return a Promise");
    return Promise.resolve();
  }

  function sameOriginWorker() {
    p = navigator.serviceWorker.register("http://some-other-origin/worker.js");
    return p.then(function(w) {
      ok(false, "Worker from different origin should fail");
    }, function(e) {
      ok(e.name === "SecurityError", "Should fail with a SecurityError");
    });
  }

  function sameOriginScope() {
    p = navigator.serviceWorker.register("worker.js", { scope: "http://www.example.com/*" });
    return p.then(function(w) {
      ok(false, "Worker controlling scope for different origin should fail");
    }, function(e) {
      ok(e.name === "SecurityError", "Should fail with a SecurityError");
    });
  }

  function httpsOnly() {
    var promise = new Promise(function(resolve) {
      SpecialPowers.pushPrefEnv({'set': [["dom.serviceWorkers.testing.enabled", false]] }, resolve);
    });

    return promise.then(function() {
      return navigator.serviceWorker.register("/worker.js");
    }).then(function(w) {
      ok(false, "non-HTTPS pages cannot register ServiceWorkers");
    }, function(e) {
      ok(e.name === "SecurityError", "Should fail with a SecurityError");
    }).then(function() {
      return new Promise((resolve) => SpecialPowers.popPrefEnv(resolve));
    });
  }

  function realWorker() {
    var p = navigator.serviceWorker.register("worker.js", { scope: "realworker*" });
    return p.then(function(w) {
      ok(w instanceof ServiceWorker, "Register a ServiceWorker");
      info(w.scope);
      ok(w.scope == (new URL("realworker*", document.baseURI)).href, "Scope should match");
      ok(w.url == (new URL("worker.js", document.baseURI)).href, "URL should be of the worker");
    }, function(e) {
      info("Error: " + e.name);
      ok(false, "realWorker Registration should have succeeded!");
    });
  }

  function abortPrevious() {
    var p = navigator.serviceWorker.register("worker2.js", { scope: "foo/*" });
    var q = navigator.serviceWorker.register("worker3.js", { scope: "foo/*" });

    return Promise.all([
      p.then(function(w) {
        ok(false, "First registration should fail with AbortError");
      }, function(e) {
        ok(e.name === "AbortError", "First registration should fail with AbortError");
      }),

      q.then(function(w) {
        ok(w instanceof ServiceWorker, "Second registration should succeed");
      }, function(e) {
        ok(false, "Second registration should succeed");
      })
    ]);
  }

  function networkError404() {
    return navigator.serviceWorker.register("404.js").then(function(w) {
        todo(false, "Should fail with NetworkError");
      }, function(e) {
        todo(e.name === "NetworkError", "Should fail with NetworkError");
      });
  }

  function parseError() {
    var p = navigator.serviceWorker.register("parse_error_worker.js");
    return p.then(function(w) {
      ok(false, "Registration should fail with parse error");
    }, function(e) {
    info("NSM " + e.name);
      ok(e instanceof Error, "Registration should fail with parse error");
    });
  }

  // FIXME(nsm): test for parse error when Update step doesn't happen (directly from register).

  function updatefound() {
    var frame = document.createElement("iframe");
    frame.setAttribute("id", "simpleregister-frame");
    frame.setAttribute("src", new URL("simpleregister/index.html", document.baseURI).href);
    document.body.appendChild(frame);
    var resolve, reject;
    var p = new Promise(function(res, rej) {
      resolve = res;
      reject = rej;
    });

    function continueTest() {
      navigator.serviceWorker.register("worker2.js", { scope: "simpleregister/*" });
    }

    window.onmessage = function(e) {
      if (e.data == "ready") {
        continueTest();
      } else if (e.data == "updatefound") {
        window.onmessage = null;
        // We have to make frame navigate away, otherwise it will call
        // MaybeStopControlling() when this document is unloaded. At that point
        // the pref has been disabled, and so MaybeStopControlling() will just
        // return since it is currently gated.
        frame.setAttribute("src", new URL("about:blank").href);
        resolve();
      }
    }
    return p;
  }

  function runTest() {
    simpleRegister()
      .then(sameOriginWorker)
      .then(sameOriginScope)
      .then(httpsOnly)
      .then(realWorker)
      .then(abortPrevious)
      .then(networkError404)
      .then(parseError)
      .then(updatefound)
      // put more tests here.
      .then(function() {
        SimpleTest.finish();
      }).catch(function(e) {
        ok(false, "Some test failed with error " + e);
        SimpleTest.finish();
      });
  }

  SimpleTest.waitForExplicitFinish();
  SpecialPowers.pushPrefEnv({"set": [
    ["dom.serviceWorkers.enabled", true],
    ["dom.serviceWorkers.testing.enabled", true]
  ]}, runTest);
</script>
</pre>
</body>
</html>
